divide some func
